feat(backend): ship phase 7 support/admin hardening

+++ src/main/java/com/vibegraph/auth/dto/AdminCreditOverviewResponse.java
@@ -7,9 +7,9 @@ import com.vibegraph.auth.domain.CreditLedger;
 
 public record AdminCreditOverviewResponse(
         UUID userId,
-        int currentCreditsLimit,
-        int creditsUsed,
-        int creditsAdjustment,
-        int creditBalance,
+        long currentCreditsLimit,
+        long creditsUsed,
+        long creditsAdjustment,
+        long creditBalance,
         List<CreditLedger> ledgerHistory
 ) {}
